requirements.txt
streamlit
openai
langchain
chromadb
sentence-transformers
python-dotenv
requests